requirements.txt
fastapi
uvicorn
asyncpg
aiokafka
pydantic
requests
pytest
python-dotenv
openai-agents
sentence-transformers
google-cloud-pubsub
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
